+++ Assets/Scripts/UserMovementNode.cs
@@ -8,7 +8,7 @@ public class UserMovementNode : PathfindingNode
     [SerializeField] public static int nbFPS;
     // Start is called before the first frame update
     public UserMovementNode(int nombre, string nom, bool estEndroitPublic, Étage étage,
-        List<int> connectedNodes, GPSCoordinate coordonéesGps) : base(nombre, nom, estEndroitPublic, étage,
+        List<float> connectedNodes, GPSCoordinate coordonéesGps) : base(nombre, nom, estEndroitPublic, étage,
         connectedNodes, coordonéesGps)
     {
     }
